+++ examples/article-excerpts/article-excerpts.ts
@@ -1,7 +1,7 @@
 /* for standalone projects replace '../../src/index' with '@get-set-fetch/scraper' */
 import path from 'path';
 import { destination } from 'pino';
-import { Scraper, PluginStore, setLogger, ScrapeEvent } from '../../src/index';
+import { PluginStore, Scraper, Project, setLogger, ScrapeEvent, CsvExporter } from '../../src/index';
 
 /* scrape configuration */
 import ScrapeConfig from './article-excerpts-config.json';
@@ -22,8 +22,9 @@ setLogger({ level: 'info' }, destination('scrape.log'));
   /* create a scraper instance with the above settings */
   const scraper = new Scraper(ScrapeConfig.storage, ScrapeConfig.client);
 
-  scraper.on(ScrapeEvent.ProjectScraped, async () => {
-    await scraper.export('article-excerpts.csv', { type: 'csv' });
+  scraper.on(ScrapeEvent.ProjectScraped, async (project: Project) => {
+    const exporter = new CsvExporter({ filepath: 'article-excerpts.csv' });
+    await exporter.export(project);
   });
 
   /* start scraping by specifying project and concurrency settings */