+++ examples/product-details/product-details.ts
@@ -1,6 +1,6 @@
 /* for standalone projects replace '../../src/index' with '@get-set-fetch/scraper' */
 import { destination } from 'pino';
-import { Scraper, setLogger, ScrapeEvent } from '../../src/index';
+import { Scraper, Project, setLogger, ScrapeEvent, CsvExporter, ZipExporter } from '../../src/index';
 
 /* scrape configuration */
 import ScrapeConfig from './product-details-config.json';
@@ -12,13 +12,15 @@ setLogger({ level: 'info' }, destination('scrape.log'));
 const scraper = new Scraper(ScrapeConfig.storage, ScrapeConfig.client);
 
 /* export books details as csv */
-scraper.on(ScrapeEvent.ProjectScraped, async () => {
-  await scraper.export('books.csv', { type: 'csv' });
+scraper.on(ScrapeEvent.ProjectScraped, async (project: Project) => {
+  const exporter = new CsvExporter({ filepath: 'books.csv' });
+  await exporter.export(project);
 });
 
 /* export book covers as zip */
-scraper.on(ScrapeEvent.ProjectScraped, async () => {
-  await scraper.export('book-covers.zip', { type: 'zip' });
+scraper.on(ScrapeEvent.ProjectScraped, async (project: Project) => {
+  const exporter = new ZipExporter({ filepath: 'book-covers.zip' });
+  await exporter.export(project);
 });
 
 /* start scraping by specifying project and concurrency settings */