+++ src/index.ts
@@ -1,14 +1,7 @@
 /* eslint-disable @typescript-eslint/no-var-requires */
 /* eslint-disable object-curly-newline */
 
-function moduleExists(name) {
-  try {
-    return require.resolve(name);
-  }
-  catch (e) {
-    return false;
-  }
-}
+import { moduleExists } from './plugins/file-utils';
 
 const KnexStorage = moduleExists('knex') ? require('./storage/knex/KnexStorage').default : null;
 const PuppeteerClient = moduleExists('puppeteer') ? require('./browserclient/PuppeteerClient').default : null;
@@ -31,13 +24,14 @@ export { default as Storage, StorageOptions } from './storage/base/Storage';
 export { initStorage } from './storage/storage-utils';
 
 export { default as Project } from './storage/base/Project';
-export { default as Resource } from './storage/base/Resource';
+export { default as Resource, ResourceQuery } from './storage/base/Resource';
 export { default as Plugin, PluginOpts } from './plugins/Plugin';
 export { default as PluginStore, StoreEntry } from './pluginstore/PluginStore';
 export { default as Scraper, ScrapeConfig, ProjectOptions, ScrapeEvent } from './scraper/Scraper';
 
 export { getLogger, setLogger } from './logger/Logger';
 export { encode, decode } from './confighash/config-hash';
+export { default as Exporter } from './export/Exporter';
 export { default as CsvExporter } from './export/CsvExporter';
 export { default as ZipExporter } from './export/ZipExporter';
 export { pipelines, mergePluginOpts } from './pipelines/pipelines';